Support ObjectID in findByFields() (#75)
The `findByFields()` method could not find documents by a given
ObjectID because the `stringToId()` function was only being run when
querying for a field named `_id`. Since any MongoDB field could in
theory contain an ObjectID, this broke queries looking for ObjectIDs in
arbitrary fields. To resolve this, remove the condition testing for the
field name of `_id` and run `stringToId()` on all values. This will
validate whether a given String value is actually in the ObjectID format
before proceeding, so it seems safe to apply everywhere.

Fixes #65